Separate classes in their own files
Change based on PR feedback
