+++ Source/Applications/Backend/applications/Tenant.ts
@@ -1,0 +1,12 @@
+// Copyright (c) Dolittle. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+import { Field } from 'type-graphql';
+import { prop } from '@typegoose/typegoose';
+
+export class Tenant {
+    @Field({ name: 'id' })
+    @prop()
+    _id?: string;
+
+    name?: string;
+}
